back-end/user_management/user_app/seializers.py
from rest_framework import serializers
from .models import UserBolgData
class UserBolgDataSerializer(serializers.ModelSerializer):
    class Meta:
        model = UserBolgData
        fields = '__all__'